APEXMALHAR-2223 ability to buffer data up to a threshold and write it to wal when the threshold is reached.
