+++ library/src/test/java/org/apache/apex/malhar/lib/state/managed/IncrementalCheckpointManagerImplTest.java
@@ -40,6 +40,7 @@ import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.fs.RemoteIterator;
 
 import com.google.common.base.Preconditions;
+import com.google.common.collect.Maps;
 
 import com.datatorrent.api.Context;
 import com.datatorrent.lib.fileaccess.FileAccessFSImpl;
@@ -47,11 +48,11 @@ import com.datatorrent.lib.util.KryoCloneUtils;
 import com.datatorrent.lib.util.TestUtils;
 import com.datatorrent.netlet.util.Slice;
 
-public class IncrementalCheckpointManagerTest
+public class IncrementalCheckpointManagerImplTest
 {
   class TestMeta extends TestWatcher
   {
-    IncrementalCheckpointManager checkpointManager;
+    IncrementalCheckpointManagerImpl checkpointManager;
     String applicationPath;
     int operatorId = 1;
     MockManagedStateContext managedStateContext;
@@ -68,7 +69,7 @@ public class IncrementalCheckpointManagerTest
       ((FileAccessFSImpl)managedStateContext.getFileAccess()).setBasePath(applicationPath + "/" + "bucket_data");
       managedStateContext.getFileAccess().init();
 
-      checkpointManager = new IncrementalCheckpointManager();
+      checkpointManager = new IncrementalCheckpointManagerImpl();
 
       managedStateContext.getTimeBucketAssigner().setup(managedStateContext);
       managedStateContext.getBucketsFileSystem().setup(managedStateContext);
@@ -89,26 +90,103 @@ public class IncrementalCheckpointManagerTest
   @Test
   public void testSerde() throws IOException
   {
-    IncrementalCheckpointManager deserialized = KryoCloneUtils.cloneObject(testMeta.checkpointManager);
+    IncrementalCheckpointManagerImpl deserialized = KryoCloneUtils.cloneObject(testMeta.checkpointManager);
     Assert.assertNotNull("state window data manager", deserialized);
   }
 
   @Test
-  public void testSave() throws IOException
+  public void testWrite() throws IOException
   {
     testMeta.checkpointManager.setup(testMeta.managedStateContext);
     Map<Long, Map<Slice, Bucket.BucketedValue>> buckets5 = ManagedStateTestUtils.getTestData(0, 5, 0);
-    testMeta.checkpointManager.save(buckets5, 10, false);
+    testMeta.checkpointManager.writeBuckets(10, buckets5);
+    testMeta.checkpointManager.beforeCheckpoint(10);
     testMeta.checkpointManager.teardown();
 
-    KryoCloneUtils<IncrementalCheckpointManager> cloneUtils = KryoCloneUtils.createCloneUtils(testMeta.checkpointManager);
+    KryoCloneUtils<IncrementalCheckpointManagerImpl> cloneUtils = KryoCloneUtils.createCloneUtils(
+        testMeta.checkpointManager);
     testMeta.checkpointManager = cloneUtils.getClone();
     testMeta.checkpointManager.setup(testMeta.managedStateContext);
-    @SuppressWarnings("unchecked")
-    Map<Long, Map<Slice, Bucket.BucketedValue>> buckets5After = (Map<Long, Map<Slice, Bucket.BucketedValue>>)
-        testMeta.checkpointManager.retrieve(10);
+
+    Map<Long, Map<Slice, Bucket.BucketedValue>> buckets5After = testMeta.checkpointManager.retrieveAllWindows().get(10L);
 
     Assert.assertEquals("saved", buckets5, buckets5After);
+    testMeta.checkpointManager.teardown();
+  }
+
+  @Test
+  public void testStateReset() throws IOException
+  {
+    KryoCloneUtils<IncrementalCheckpointManagerImpl> cloneUtils = KryoCloneUtils.createCloneUtils(
+        testMeta.checkpointManager);
+    IncrementalCheckpointManagerImpl checkpointManagerCheckpointed = cloneUtils.getClone();
+
+    testMeta.checkpointManager.setup(testMeta.managedStateContext);
+    Map<Long, Map<Slice, Bucket.BucketedValue>> buckets5 = ManagedStateTestUtils.getTestData(0, 5, 0);
+    testMeta.checkpointManager.writeBuckets(10, buckets5);
+    testMeta.checkpointManager.beforeCheckpoint(10);
+    testMeta.checkpointManager.teardown();
+
+    testMeta.checkpointManager = checkpointManagerCheckpointed;
+    testMeta.checkpointManager.setup(testMeta.managedStateContext);
+
+    Assert.assertTrue("State should be reset as there wasn't any thing checkpointed",
+        testMeta.checkpointManager.retrieveAllWindows().isEmpty());
+    testMeta.checkpointManager.teardown();
+  }
+
+  @Test
+  public void testMultipleWritesInAWindow() throws IOException
+  {
+    testMeta.checkpointManager.setup(testMeta.managedStateContext);
+    Map<Long, Map<Slice, Bucket.BucketedValue>> buckets5 = ManagedStateTestUtils.getTestData(0, 5, 0);
+    testMeta.checkpointManager.writeBuckets(10, buckets5);
+
+    Map<Long, Map<Slice, Bucket.BucketedValue>> buckets10 = ManagedStateTestUtils.getTestData(5, 10, 0);
+    testMeta.checkpointManager.writeBuckets(10, buckets10);
+    testMeta.checkpointManager.beforeCheckpoint(10);
+    testMeta.checkpointManager.teardown();
+
+    KryoCloneUtils<IncrementalCheckpointManagerImpl> cloneUtils = KryoCloneUtils.createCloneUtils(
+        testMeta.checkpointManager);
+    testMeta.checkpointManager = cloneUtils.getClone();
+    testMeta.checkpointManager.setup(testMeta.managedStateContext);
+
+    Map<Long, Map<Long, Map<Slice, Bucket.BucketedValue>>> recovered = testMeta.checkpointManager.retrieveAllWindows();
+
+    Map<Long, Map<Long, Map<Slice, Bucket.BucketedValue>>> allBuckets10 = Maps.newHashMap();
+    IncrementalCheckpointManagerImpl.addAdditionalBucketDataForWindow(10, allBuckets10, buckets5);
+    IncrementalCheckpointManagerImpl.addAdditionalBucketDataForWindow(10, allBuckets10, buckets10);
+
+    Assert.assertEquals("saved", allBuckets10, recovered);
+    testMeta.checkpointManager.teardown();
+  }
+
+  @Test
+  public void testMultipleWritesInAWindowWithBucketsOverlap() throws IOException
+  {
+    testMeta.checkpointManager.setup(testMeta.managedStateContext);
+    Map<Long, Map<Slice, Bucket.BucketedValue>> buckets5 = ManagedStateTestUtils.getTestData(0, 5, 0);
+    testMeta.checkpointManager.writeBuckets(10, buckets5);
+
+    Map<Long, Map<Slice, Bucket.BucketedValue>> buckets7 = ManagedStateTestUtils.getTestData(2, 7, 0);
+    testMeta.checkpointManager.writeBuckets(10, buckets7);
+    testMeta.checkpointManager.beforeCheckpoint(10);
+    testMeta.checkpointManager.teardown();
+
+    KryoCloneUtils<IncrementalCheckpointManagerImpl> cloneUtils = KryoCloneUtils.createCloneUtils(
+        testMeta.checkpointManager);
+    testMeta.checkpointManager = cloneUtils.getClone();
+    testMeta.checkpointManager.setup(testMeta.managedStateContext);
+
+    Map<Long, Map<Long, Map<Slice, Bucket.BucketedValue>>> recovered = testMeta.checkpointManager.retrieveAllWindows();
+
+    Map<Long, Map<Long, Map<Slice, Bucket.BucketedValue>>> allBuckets7 = Maps.newHashMap();
+    IncrementalCheckpointManagerImpl.addAdditionalBucketDataForWindow(10, allBuckets7, buckets5);
+    IncrementalCheckpointManagerImpl.addAdditionalBucketDataForWindow(10, allBuckets7, buckets7);
+
+    Assert.assertEquals("num buckets", 7, allBuckets7.get(10L).size());
+    Assert.assertEquals("saved", allBuckets7, recovered);
     testMeta.checkpointManager.teardown();
   }
 
@@ -118,7 +196,7 @@ public class IncrementalCheckpointManagerTest
     testMeta.checkpointManager.setup(testMeta.managedStateContext);
 
     Map<Long, Map<Slice, Bucket.BucketedValue>> buckets5 = ManagedStateTestUtils.getTestData(0, 5, 0);
-    testMeta.checkpointManager.save(buckets5, 10, false);
+    testMeta.checkpointManager.writeBuckets(10, buckets5);
     //Need to synchronously call transfer window files so shutting down the other thread.
     testMeta.checkpointManager.teardown();
     Thread.sleep(500);
@@ -144,7 +222,7 @@ public class IncrementalCheckpointManagerTest
     testMeta.checkpointManager.setup(testMeta.managedStateContext);
 
     Map<Long, Map<Slice, Bucket.BucketedValue>> data = ManagedStateTestUtils.getTestData(0, 5, 0);
-    testMeta.checkpointManager.save(data, 10, false);
+    testMeta.checkpointManager.writeBuckets(10, data);
     testMeta.checkpointManager.committed(10);
     latch.await();
     testMeta.checkpointManager.teardown();
@@ -194,5 +272,5 @@ public class IncrementalCheckpointManagerTest
     }
   }
 
-  private static final transient Logger LOG = LoggerFactory.getLogger(IncrementalCheckpointManagerTest.class);
+  private static final transient Logger LOG = LoggerFactory.getLogger(IncrementalCheckpointManagerImplTest.class);
 }
